fix bug when color placement count is negative

+++ Arduino/libraries/Buffer/buffer.h
@@ -23,7 +23,7 @@ class Buffer
   void display_buffer(rgb_color * pbuffer);
   void render_display();
   void erase(bool display);
-  void push_color(rgb_color color, byte times, bool display, byte effect, byte max, byte start, bool color_correction);
+  void push_color(rgb_color color, int times, bool display, byte effect, byte max, byte start, bool color_correction);
   void push_rgb_color(byte red, byte green, byte blue);
   void push_hsl_color(int hue, int sat, int lit);
   void push_carry_color();
@@ -236,7 +236,7 @@ void Buffer::shift_buffer(rgb_color * buffer, byte * effects, byte max, byte sta
 }
 
 // todo: only need to shift buffer multiple times if displaying (otherwise, shift the whole amount)
-void Buffer::push_color(rgb_color color, byte times = 1, bool display = false, byte effect = DEFAULT_EFFECT, byte max = 0, byte start = 0, bool color_correction = true)
+void Buffer::push_color(rgb_color color, int times = 1, bool display = false, byte effect = DEFAULT_EFFECT, byte max = 0, byte start = 0, bool color_correction = true)
 {
   rgb_color * buffer = buffers[current_display];
   byte * effects = effects_buffers[current_display];
